+++ QSVPipeline/rgy_input_raw.h
@@ -63,7 +63,7 @@ public:
     virtual rgy_rational<int> getInputTimebase() const override {
         if (m_timebase.is_valid()) return m_timebase;
         return (m_inputVideoInfo.type == RGY_INPUT_FMT_Y4M)
-            ? rgy_rational<int>(1, 1000000) : RGYInput::getInputTimebase();
+            ? rgy_rational<int>(1, 1200000) : RGYInput::getInputTimebase();
     }
     virtual int64_t GetVideoFirstKeyPts() const override;
 
@@ -79,9 +79,7 @@ protected:
     bool m_isPipe;
     RGYParamParallelEncPipeHandle m_chunkPipeHandle;
     int64_t m_firstKeyPts;
-    int64_t m_y4mTimestamp;
     int64_t m_y4mPreviousTimestamp;
-    bool m_y4mHasTimestamp;
     bool m_y4mTimestampWarningShown;
 };